commit pipes modification

+++ x64barebones/Kernel/include/pipeMaster.h
@@ -22,12 +22,12 @@ void createPipeMaster();
 PipeMasterADT getPipeMaster();
 PipeADT createPipe(uint64_t id);
 Node * getPipeNodeById(uint64_t id);
-uint16_t pipeOpen(uint64_t id, char mode);
-uint16_t pipeOpenAnonymous(int id, char mode, uint32_t pid);
-uint16_t pipeClose(uint64_t id);
-uint16_t pipeCloseAnonymous(int id, uint32_t pid);
-uint16_t pipeWrite(uint64_t id, uint32_t pid, char* msg, int len);
-uint16_t pipeRead(uint64_t id, uint32_t pid, char* buffer, int len, uint32_t * readBytes);
-uint64_t getPipeId();
+int16_t pipeOpen(int id, char mode);
+int16_t pipeOpenAnonymous(int id, char mode, uint32_t pid);
+int16_t pipeClose(int id);
+int16_t pipeCloseAnonymous(int id, uint32_t pid);
+int16_t pipeWrite(uint64_t id, uint32_t pid, char* msg, int len);
+int16_t pipeRead(uint64_t id, uint32_t pid, char* buffer, int len, uint32_t * readBytes);
+int64_t getPipeId();
 
 #endif
